Partitioned Table Fix Estimated Tuple Count On Base Table
As mentioned in your recent live stream, due to the -1 estimated tuple count on base tables, the rolled up partitioned table values contains a different count relative to the indexes which all begin with an estimated count of 0. You proposed fixing it and then went on without addressing it and I didn't want this to get lost.