Pulling the ordinary tables when the partitioned table is queried for
When using check_indexes on a partitioned table, it is a little confusing that you can't get the underlying tables
at the same time unless you ask for null to get everything. While this might need to be an additional option in the
future, right now it seems to make the most sense to return the ordinary tables with the partitioned table since
they are all really one thing. There are also implications of Vacuum/Analyze dates and the warnings that may
appear on the underlying ordinary table that are suppressed if you only look at the partitioned table level.