Fix last_vacuum column removed in error from Child Partitioned Tables
During the merge there was a fix on the error of bad column name last_vacuum,
the problem was you renamed it to last_manual_nonfull_vacuum instead of last_vacuum to be more precise.
I have added the column back into the child partitioned tables so that you can see the effect.
To test, run a manual vacuum on an underlying partition table such as:

    vacuum users_partitioned_2020;
Before this fix is in place, the last_manual_nonfull_vacuum column will continue to show null.
After the fix that column should show the timestamp appropriatly.