Added dead_tuples to Child Partitioned Tables
You had the logic built out to put the dead tuple count on the main ordinary tables.
This adds that logic to the child ordinary tables of a partitioned tables.
To test this, first do a delete on a child partitioned tables:

    DELETE FROM users_partitioned_2020
    WHERE id % 10 = 0;
Once that delete is done, check_indexes will show it on the partitioned table but not the ordinary table.
After this fix it will appear on the ordinary table as well.